backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
Pillow>=10.0.0
boto3>=1.28.0
python-multipart>=0.0.6
